Commonize _nDefaultFallbackChar and relevant tests (skip on JS though)
